KAN-41 : Rename: move package
