+++ src/main/java/capstone/relation/websocket/stomp/StompPreHandler.java
@@ -1,4 +1,4 @@
-package capstone.relation.websocket;
+package capstone.relation.websocket.stomp;
 
 import java.time.Instant;
