+++ src/main/java/capstone/relation/websocket/stomp/WebSocketConfig.java
@@ -1,4 +1,4 @@
-package capstone.relation.websocket;
+package capstone.relation.websocket.stomp;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.messaging.simp.config.ChannelRegistration;
@@ -7,6 +7,7 @@ import org.springframework.web.socket.config.annotation.EnableWebSocketMessageBr
 import org.springframework.web.socket.config.annotation.StompEndpointRegistry;
 import org.springframework.web.socket.config.annotation.WebSocketMessageBrokerConfigurer;
 
+import capstone.relation.websocket.StompHandshakeHandler;
 import lombok.RequiredArgsConstructor;
 
 @Configuration
